+++ common/src/main/java/earth/terrarium/adastra/mixins/common/radio/ServerPlayerMixin.java
@@ -1,8 +1,8 @@
 package earth.terrarium.adastra.mixins.common.radio;
 
-import earth.terrarium.adastra.common.utils.radio.RadioHolder;
 import earth.terrarium.adastra.common.network.NetworkHandler;
 import earth.terrarium.adastra.common.network.messages.ClientboundPlayStationPacket;
+import earth.terrarium.adastra.common.utils.radio.RadioHolder;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.player.Player;